+++ custom_components/ezviz_plus/entity.py
@@ -94,7 +94,10 @@ class EzvizEntity(CoordinatorEntity[EzvizDataUpdateCoordinator], Entity):
     @property
     def available(self) -> bool:
         """Return True if entity is available."""
-        return bool(self.data["status"] != 2)
+        return bool(
+            self.coordinator.is_camera_available(self._serial)
+            and self.data["status"] != 2
+        )
 
 
 class EzvizBaseEntity(Entity):
@@ -122,4 +125,7 @@ class EzvizBaseEntity(Entity):
     @property
     def available(self) -> bool:
         """Return True if entity is available."""
-        return bool(self.data["status"] != 2)
+        return bool(
+            self.coordinator.is_camera_available(self._serial)
+            and self.data["status"] != 2
+        )